fix(orders): 테스트 DB 오염 버그 수정 + 실제 주문내역 동기화 추가 (#38)
* fix(orders): 테스트가 프로덕션 DB를 오염시키던 버그 수정 + 실제 주문내역 동기화 추가

## 근본 원인

tests/integration/test_toss_roundtrip.py의 test_signal_to_toss_order_roundtrip이
StateStore()를 인자 없이 생성했는데, 기본 경로가 프로덕션 DB
(~/quanteo/data/quanteo.db)를 가리킨다. MockTossRestClient(실제 Toss API를
호출하지 않는 가짜 클라이언트)로 executor.submit()을 호출하는 이 테스트가
pytest를 돌릴 때마다 가짜 'submitted' 주문을 실제 대시보드 주문내역에
그대로 삽입해왔다 (재현: 테스트 1회 실행 → orders 테이블 5행 → 6행).

## 수정

- test_signal_to_toss_order_roundtrip이 StateStore(":memory:")를 쓰도록 수정
- tests/conftest.py에 세션 전체 안전장치 추가 — 어떤 테스트든 프로덕션
  orders 테이블을 건드리면 즉시 실패시켜 같은 사고 재발을 원천 차단
- 기존에 쌓인 가짜 주문 6건은 DB에서 직접 정리 (백업 후 삭제)

## 추가 조사에서 드러난 구조적 문제

로컬 orders 테이블은 OrderExecutor.submit()이 만든 행만 반영하도록
설계됐는데, 이 submit()을 실제로 호출하는 신호 파이프라인이 없다
(StrategyEngine이 SIGNAL을 발행해도 RiskManager.evaluate()·
executor.submit()을 호출하는 구독자가 전혀 없음). Toss는 모의투자
구분이 없는 단일 계좌라 사용자가 Toss 앱에서 직접 낸 주문도 실제
계좌 상태의 일부인데, 로컬 DB만으로는 이를 반영할 수 없었다.

- core/execution/order_history_sync.py (신규): Toss list_orders()를
  브로커 source of truth로 삼아 orders 테이블 전체를 주기적으로
  동기화하는 OrderHistorySyncFeed 추가. 최초 1회 CLOSED 이력을 커서
  끝까지 페이지네이션해 백필하고, 이후엔 최신 페이지만 확인
  (ORDER_HISTORY Rate Limit 절약). --with-trading 여부와 무관하게
  항상 동작 (읽기 전용 GET만 수행)
- core/store/db.py: upsert_broker_order() 추가 — clientOrderId가 있으면
  기존 행 갱신, 없으면(Toss 앱에서 직접 낸 주문) toss-native-{orderId}로
  새로 삽입
- core/app.py: OrderHistorySyncFeed를 PositionSyncFeed와 같은 자리에 배선

## 백필 중 추가로 발견한 크래시

실제 운영 DB로 백필을 검증하다가 해외주식 fractional investing 체결
수량("0.000151")을 만나 _safe_int()가 RuntimeError로 죽는 걸 발견 —
positions.qty는 이미 REAL로 이 문제를 처리해뒀는데 orders.qty와
TossOrder.quantity/OrderExecution.filled_quantity는 int로 남아있던
누락이었다. 같은 패턴으로 float/REAL로 맞춰 수정 (회귀 테스트 포함).

실제 계정 데이터로 검증 결과 총 327건(cancelled 90·filled 211·
pending 3·rejected 23)의 실제 주문 이력이 정상적으로 반영됨을 확인.

* fix(orders): 리뷰에서 지적된 백필 완료 오판·소수점 truncation·stale 주석 수정

## 수정

- OrderHistorySyncFeed._fetch_all_closed_orders()가 _MAX_BACKFILL_PAGES
  상한에 걸려 일부만 받아도 무조건 백필 완료(True)로 표시하던 버그 수정.
  이제 완료 여부를 함께 반환해, 상한에 걸리면 다음 주기에도 전체 백필을
  다시 시도한다 (대량 거래 계정에서 오래된 이력이 영원히 누락되는 문제 방지)
- get_pending_orders()가 qty를 int()로 잘라 해외 fractional 주문이
  재시작 복구 로그에서 0으로 보이던 문제 수정 (PendingOrder.qty: int → float,
  db.py의 캐스팅과 app.py의 %d 포맷 문자열도 함께 수정)
- api/models.py의 PositionItem docstring이 "OrderItem.qty는 항상 정수"라고
  주장하던 stale 주석 수정 (이전 커밋에서 실제로는 float로 바꿔놓고 주석은
  안 고쳤던 자기모순)

## 테스트 추가

- upsert_broker_order()가 broker_order_id만으로 기존 행을 찾아 갱신하는
  경로(clientOrderId가 나중에 채워지는 시나리오) 테스트
- upsert_broker_order()를 통한 소수점 수량 저장 round-trip 테스트
- 백필이 페이지 상한에 걸렸을 때 다음 주기에 전체 백필을 재시도하는지
  검증하는 회귀 테스트

전체 729개 테스트 통과(기존부터 있던 무관한 테스트 1개 제외), 실제 서버
재시작으로 재시작 복구 로그에 소수점 수량이 정확히 표시되는 것까지 확인.

---------

+++ core/adapters/toss/models.py
@@ -58,7 +58,10 @@ class Commission:
 class OrderExecution:
     """주문 체결 정보 (Order.execution 내부 필드)."""
 
-    filled_quantity: int
+    # float: 해외주식(미국)은 소수점 단위 매매(fractional investing)를 지원해
+    # 체결 수량이 정수가 아닐 수 있다(예: "0.000151"). int로 두면 Toss가
+    # 이런 값을 내려줄 때 _parse_toss_order()에서 파싱 자체가 실패한다.
+    filled_quantity: float
     avg_fill_price: Decimal | None
     fees: Decimal | None
 
@@ -86,7 +89,7 @@ class TossOrder:
     side: Literal["BUY", "SELL"]
     order_type: Literal["LIMIT", "MARKET"]
     status: str
-    quantity: int
+    quantity: float  # 해외 fractional investing 지원 — filled_quantity와 동일한 이유
     currency: str
     ordered_at: datetime
     execution: OrderExecution
